backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
passlib[bcrypt]
pyjwt